+++ Makefile
@@ -4,7 +4,7 @@ HOST ?= 127.0.0.1
 PORT ?= 8000
 
 dev:
-	uvicorn app:app --reload --host $(HOST) --port $(PORT)
+	uvicorn ditherbooth.app:app --reload --host $(HOST) --port $(PORT)
 
 format:
 	black .
