+++ ditherbooth/imaging/process.py
