+++ ditherbooth/static/examples/final_1bit_463.png
